API_project/requirements.txt
djangopip
djangorestframework